[fix] don't clean build and output directories when using `--watch` (#7516)
* Dont clean build and output_dir when watching

When watching static directory is missing on outDir
because the vite svelte-kit plugin clean them

* After nnpm changeset

* Update .changeset/large-starfishes-relate.md

* Update packages/kit/src/exports/vite/index.js

* Update packages/kit/src/exports/vite/index.js

+++ packages/kit/src/exports/vite/index.js
@@ -345,10 +345,11 @@ function kit() {
 			completed_build = false;
 
 			if (is_build) {
-				rimraf(paths.build_dir);
+				if (!vite_config.build.watch) {
+					rimraf(paths.build_dir);
+					rimraf(paths.output_dir);
+				}
 				mkdirp(paths.build_dir);
-
-				rimraf(paths.output_dir);
 				mkdirp(paths.output_dir);
 			}
 		},
